Numark Mixtrack Platinum: add mapping
This is a mapping for the Numark Mixtrack Platinum. Most features of
the controller are supported including the LCD screens and 4 deck
control. This mapping is targeted for Mixxx 2.0.0.

Notes:
- FX buttons always operat channels 1 and 2. This is a limitation of
the controller that could be scripted around, but I got lazy (also how
it works in Serato).
- Both FX dry/wet strips control the same effects chain and will affect
both channels 1 and 2 if effects are enabled on them. Serato has
isolated effects banks for each channel.
- The "BEATS" knobs are not mapped, this concept did not translate
cleanly to Mixxx.
- Sometimes there are artifacts in the auto after long spinback
maneuvers. This may not have anything to do with the controller.
- The sub-beat auto-loop buttons do not light up when enabled unless
the shift key is pressed. This is different from how the work in Serato.
- Master sync enable and disable works like in the Mixxx software interface,
not how it works in Serato. Shift-sync is not mapped.
- The slip mode indicator is not mapped, this is trivial to fix (midi
codes are documented in the JS file), but I couldn't figure out how to
use slip mode in the Mixxx interface.
- The up/down indicators on the LCD screens to indicate which direction
to adjust the pitch slider are not implemented. It's possible to do, but
would require some scripting.
- Browse knob press is unmapped. I couldn't get this to do something
sensible in Mixxx 2.0.0. Version 2.1.0 seems to have more sensible
options.